# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Beat Portal API
  description: |
    FastAPI backend for Beat Portal - A DJ music library organizer with AI-powered metadata enhancement.
    
    ## Features
    - Offline-first music library management
    - Advanced metadata analysis (BPM, Key, Genre, Mood)
    - AI-powered metadata enhancement
    - Track organization and playlist management
  version: 1.0.0
  contact:
    name: Beat Portal Support
    email: support@beatportal.dev
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: http://localhost:8000
    description: Local base server
...
paths:
...
  /metadata/analyze:
    post:
      tags:
        - Metadata
      summary: Analyze Track Metadata
      description: Analyze audio file to extract BPM, key, and other metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeMetadataRequest'
      responses:
        '200':
          description: Metadata analyzed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataAnalysis'
  /metadata/batch-analyze:
    post:
      tags:
        - Metadata
      summary: Batch Analyze Metadata
      description: Analyze metadata for multiple tracks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchAnalyzeMetadataRequest'
      responses:
        '202':
          description: Batch analysis started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchAnalyzeMetadataResponse'
...
components:
  schemas:
...
    DetectedMetadata:
      type: object
      properties:
        bpm:
          type: integer
        key:
          type: string
        duration_seconds:
          type: integer
          description: Duration in seconds
        sample_rate_hz:
          type: integer
          description: Sample rate in Hertz (Hz)
        bitrate_bps:
          type: integer
          description: Bitrate in bits per second
    ConfidenceScores:
      type: object
      properties:
        bpm:
          type: number
          format: float
        key:
          type: number
          format: float
    MetadataAnalysis:
      type: object
      properties:
        file_path:
          type: string
        detected_metadata:
          $ref: '#/components/schemas/DetectedMetadata'
        confidence_scores:
          $ref: '#/components/schemas/ConfidenceScores'
...
    AnalyzeMetadataRequest:
      type: object
      properties:
        file_path:
          type: string
          example: /path/to/track.mp3
        analysis_options:
          type: object
          properties:
            detect_bpm:
              type: boolean
              default: true
            detect_key:
              type: boolean
              default: true
            detect_genre:
              type: boolean
              default: false
    BatchAnalyzeMetadataRequest:
      type: object
      properties:
        track_ids:
          type: array
          items:
            type: string
            format: uuid
    BatchAnalyzeMetadataResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          example: processing
        errors:
          type: array
          items:
            type: string
          description: List of errors encountered during the batch analysis
...